pdf_editor/requirements.txt
pypdf>=3.0.0
PyMuPDF>=1.18.0
Pillow>=9.0.0